+++ test/mysql_test.go
@@ -1,7 +1,6 @@
 package test
 
 import (
-	"encoding/json"
 	"fmt"
 	"log"
 	"testing"
@@ -39,24 +38,33 @@ func TestMySQLMigration(t *testing.T) {
 	// log.Println("MYSQL MIGRATION")
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.Migrate(new(User)); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 }
 
 func TestMySQLTableExists(t *testing.T) {
 	if isExist := my.Table("User").Exists(); isExist != true {
-		log.Fatal(fmt.Errorf("Unexpected error, table %q should exists", "User"))
+		t.Fatal(fmt.Errorf("Unexpected error, table %q should exists", "User"))
+	}
+}
+
+func TestMySQLTruncate(t *testing.T) {
+	// log.Println(strings.Repeat("-", 100))
+	// log.Println("MYSQL TRUNCATE")
+	// log.Println(strings.Repeat("-", 100))
+	if err := my.Truncate(new(User)); err != nil {
+		t.Fatal(err)
 	}
 }
 
 func TestMySQLAddIndex(t *testing.T) {
 	if err := my.Table("User").
 		AddUniqueIndex("Username"); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 	if err := my.Table("User").
 		AddIndex("Age"); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 }
 
@@ -66,7 +74,7 @@ func TestMySQLCreate(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	u := getFakeUser()
 	if err := my.Create(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
@@ -74,7 +82,7 @@ func TestMySQLCreate(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	u = getFakeUser()
 	if err := my.Create(u, nameKey); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
@@ -82,7 +90,7 @@ func TestMySQLCreate(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	u = getFakeUser()
 	if err := my.Create(u, idKey); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
@@ -90,7 +98,7 @@ func TestMySQLCreate(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	uu := []User{*getFakeUser(), *getFakeUser()}
 	if err := my.Create(&uu); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
@@ -98,22 +106,9 @@ func TestMySQLCreate(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	users := []*User{getFakeUser(), getFakeUser()}
 	if err := my.Create(&users); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
-}
-
-func TestJSONMarshal(t *testing.T) {
-	u := new(User)
-	if err := my.Select("*", "Name").First(u); err != nil {
-		log.Fatal(err)
-	}
-
-	b, err := json.Marshal(u)
-	if err != nil {
-		log.Fatal(err)
-	}
-	log.Println(string(b))
 }
 
 func TestMySQLSelect(t *testing.T) {
@@ -122,7 +117,7 @@ func TestMySQLSelect(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	u := new(User)
 	if err := my.Select("*", "Name").First(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 }
 
@@ -133,7 +128,7 @@ func TestMySQLDistinctOn(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.NewQuery().
 		DistinctOn("*").First(u); err == nil {
-		log.Fatal("Expected `DistinctOn` cannot allow *")
+		t.Fatal("Expected `DistinctOn` cannot allow *")
 	}
 
 	// log.Println(strings.Repeat("-", 100))
@@ -141,7 +136,7 @@ func TestMySQLDistinctOn(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.NewQuery().
 		DistinctOn("").First(u); err == nil {
-		log.Fatal("Expected `DistinctOn` cannot have empty")
+		t.Fatal("Expected `DistinctOn` cannot have empty")
 	}
 
 	// log.Println(strings.Repeat("-", 100))
@@ -149,26 +144,26 @@ func TestMySQLDistinctOn(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.NewQuery().
 		DistinctOn("Name", "Password").First(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 }
 
 func TestMySQLEmptySliceInJSON(t *testing.T) {
 	u := new(User)
 	if err := my.First(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 	if u.Email == nil {
-		log.Fatal(fmt.Errorf("empty slice should init on any `Get` func"))
+		t.Fatal(fmt.Errorf("empty slice should init on any `Get` func"))
 	}
 
 	u2 := getFakeUser()
 	u2.Email = nil
 	if err := my.Create(u2); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 	if u2.Email == nil {
-		log.Fatal(fmt.Errorf("empty slice should init on any `Create` func"))
+		t.Fatal(fmt.Errorf("empty slice should init on any `Create` func"))
 	}
 }
 
@@ -179,47 +174,163 @@ func TestMySQLGet(t *testing.T) {
 	// log.Println("MYSQL FIRST")
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.First(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
 	// log.Println("MYSQL FIND")
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.Find(u.Key, u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
 	// log.Println("MYSQL GET")
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.Get(users); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
 	// log.Println("MYSQL GET WITH UNSCOPED")
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.NewQuery().Unscoped().Get(users); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 }
 
 func TestMySQLJSONEqual(t *testing.T) {
+	var emptyStr string
+
 	users := new([]User)
 	if err := my.NewQuery().
-		WhereJSONEqual("Address > PostCode", 85).
+		WhereJSONEqual("Address>PostCode", int32(85)).
 		Get(users); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
+	}
+
+	if err := my.NewQuery().
+		WhereJSONEqual("Address>PostCode", uint32(85)).
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+
+	postCode := uint32(85)
+	if err := my.NewQuery().
+		WhereJSONEqual("Address>PostCode", &postCode).
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+
+	if err := my.NewQuery().
+		WhereJSONEqual("Address>Line1", "7812, Jalan Section 22").
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+	if len(*users) <= 0 {
+		t.Fatal("JSON equal has unexpected result")
+	}
+
+	if err := my.NewQuery().
+		WhereJSONEqual("Address>Line2", emptyStr).
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+	if len(*users) <= 0 {
+		t.Fatal("JSON equal has unexpected result")
+	}
+
+	timeZone := new(time.Time)
+	if err := my.NewQuery().
+		WhereJSONEqual("Address>region.TimeZone", timeZone).
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+	if len(*users) <= 0 {
+		t.Fatal("JSON equal has unexpected result")
+	}
+}
+
+func TestMySQLJSONNotEqual(t *testing.T) {
+	var timeZone *time.Time
+	users := new([]User)
+	if err := my.NewQuery().
+		WhereJSONNotEqual("Address>region.TimeZone", timeZone).
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+	if len(*users) <= 0 {
+		t.Fatal("JSON equal has unexpected result")
+	}
+
+	if err := my.NewQuery().
+		WhereJSONNotEqual("Address>Country", "").
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+	if len(*users) > 0 {
+		t.Fatal("JSON equal has unexpected result")
 	}
 }
 
 func TestMySQLJSONIn(t *testing.T) {
 	users := new([]User)
 	if err := my.NewQuery().
-		WhereJSONEqual("Address > PostCode", 85).
+		WhereJSONContainAny("Address>PostCode", []interface{}{0, 10, 20}).
 		Get(users); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
+	}
+	if len(*users) <= 0 {
+		t.Fatal("JSON contain any has unexpected result")
+	}
+}
+
+func TestMySQLJSONNotIn(t *testing.T) {
+	users := new([]User)
+	if err := my.NewQuery().
+		WhereJSONNotIn("Address>Line1", []interface{}{"PJ", "KL", "Cheras"}).
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+	if len(*users) <= 0 {
+		t.Fatal("JSON contain any has unexpected result")
+	}
+}
+
+func TestMySQLJSONType(t *testing.T) {
+	users := new([]User)
+	if err := my.NewQuery().
+		WhereJSONType("Address>region", "OBJECT").
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+	if len(*users) <= 0 {
+		t.Fatal("JSON isObject has unexpected result")
+	}
+}
+
+func TestMySQLJSONIsObject(t *testing.T) {
+	users := new([]User)
+	if err := my.NewQuery().
+		WhereJSONIsObject("Address>region").
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+	if len(*users) <= 0 {
+		t.Fatal("JSON isObject has unexpected result")
+	}
+}
+
+func TestMySQLJSONIsArray(t *testing.T) {
+	users := new([]User)
+	if err := my.NewQuery().
+		WhereJSONIsArray("Address>region.keys").
+		Get(users); err != nil {
+		t.Fatal(err)
+	}
+	if len(*users) <= 0 {
+		t.Fatal("JSON isArray has unexpected result")
 	}
 }
 
@@ -232,7 +343,7 @@ func TestMySQLPaginate(t *testing.T) {
 		Limit: 10,
 	}
 	if err := my.Paginate(p, users); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 	// log.Println("Records :", p.Count())
 	// log.Println("Cursor :", p.NextCursor())
@@ -244,21 +355,21 @@ func TestMySQLUpsert(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	u := getFakeUser()
 	if err := my.Upsert(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
 	// log.Println("MYSQL SINGLE UPSERT WITH PARENT KEY (ID KEY)")
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.Upsert(u, idKey); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
 	// log.Println("MYSQL SINGLE UPSERT WITH PARENT KEY (NAME KEY)")
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.Upsert(u, nameKey); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
@@ -266,7 +377,7 @@ func TestMySQLUpsert(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	users := []*User{getFakeUser(), getFakeUser()}
 	if err := my.Upsert(&users); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
@@ -274,7 +385,7 @@ func TestMySQLUpsert(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	uu := []User{*getFakeUser(), *getFakeUser()}
 	if err := my.Upsert(&uu); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	uuu := []User{*getFakeUser(), *getFakeUser()}
@@ -282,14 +393,14 @@ func TestMySQLUpsert(t *testing.T) {
 	// log.Println("MYSQL MULTI UPSERT WITH SLICE STRUCT AND PARENT KEY (ID KEY)")
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.Upsert(&uuu, idKey); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// log.Println(strings.Repeat("-", 100))
 	// log.Println("MYSQL MULTI UPSERT WITH SLICE STRUCT AND PARENT KEY (NAME KEY)")
 	// log.Println(strings.Repeat("-", 100))
 	if err := my.Upsert(&uuu, nameKey); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 }
@@ -303,14 +414,14 @@ func TestMySQLUpdate(t *testing.T) {
 		Update(map[string]interface{}{
 			"Name": "sianloong",
 		}); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	if err := my.Table("User").Limit(1).
 		Update(map[string]interface{}{
 			"Email": []string{"abc@gmail.com", "abc@hotmail.com", "abc@yahoo.com"},
 		}); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	// TODO: support struct
@@ -318,7 +429,7 @@ func TestMySQLUpdate(t *testing.T) {
 	// 	Update(map[string]interface{}{
 	// 		"Address": Address{"", "Line2", "", 63000},
 	// 	}); err != nil {
-	// 	log.Fatal(err)
+	// 	t.Fatal(err)
 	// }
 }
 func TestMySQLSoftDelete(t *testing.T) {
@@ -327,10 +438,10 @@ func TestMySQLSoftDelete(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	u := getFakeUser()
 	if err := my.Create(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 	if err := my.Delete(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 }
 
@@ -340,10 +451,10 @@ func TestMySQLHardDelete(t *testing.T) {
 	// log.Println(strings.Repeat("-", 100))
 	u := new(User)
 	if err := my.First(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 	if err := my.Destroy(u); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 }
 
@@ -352,19 +463,19 @@ func TestMySQLTable(t *testing.T) {
 	if err := my.Table("User").
 		WhereLike("Name", "nick%").
 		Get(users); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	if err := my.Table("User").
 		Where("Age", ">", 0).
 		Get(users); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 
 	user := new(User)
 	if err := my.Table("User").
 		First(user); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 }
 
@@ -383,7 +494,7 @@ func TestMySQLRunInTransaction(t *testing.T) {
 		u.UpdatedDateTime = time.Now().UTC()
 		return txn.Save(u)
 	}); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 }
 
@@ -394,22 +505,13 @@ func TestMySQLScan(t *testing.T) {
 	var count, sum uint
 	if err := my.Table("User").Select("COALESCE(COUNT(*),0), COALESCE(SUM(Age),0)").
 		Scan(&count, &sum); err != nil {
-		log.Fatal(err)
+		t.Fatal(err)
 	}
 	log.Println("Count :", count, ", Sum :", sum)
 }
 
-func TestMySQLTruncate(t *testing.T) {
-	// log.Println(strings.Repeat("-", 100))
-	// log.Println("MYSQL TRUNCATE")
-	// log.Println(strings.Repeat("-", 100))
-	// if err := my.Truncate(new(User)); err != nil {
-	// 	log.Fatal(err)
-	// }
-}
-
 func TestMySQLDropTableIfExists(t *testing.T) {
 	// if err := my.Table("User").DropIfExists(); err != nil {
-	// 	log.Fatal(err)
+	// 	t.Fatal(err)
 	// }
 }